pyproject.toml
[project]
name = "kipsoiwet-bot"
version = "0.1.0"
description = "Telegram assistant for a Polymarket martingale workflow"
requires-python = ">=3.11"
dependencies = [
  "python-telegram-bot>=21.0,<22.0",
  "pydantic>=2.7,<3.0",
  "python-dotenv>=1.0,<2.0",
  "httpx>=0.27,<0.29"
]

[project.optional-dependencies]
dev = ["pytest>=8.2,<9.0"]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
pythonpath = ["src"]
